pyproject.toml

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SiteToSheet"
version = "0.1.0"
description = "A tool to scrape websites and update Google Sheets"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.7"
dependencies = [
    'beautifulsoup4==4.12.3',
    'googlemaps==4.10.0',
    'gspread==6.1.2',
    'numpy==2.0.1',
    'pandas==2.2.2',
    'protobuf==5.27.3',
    'python-dotenv==1.0.1',
    'ratelimit==2.2.1',
    'requests==2.32.3',
    'setuptools==69.2.0',
    'spacy==3.7.5'
]

[project.scripts]
sitetosheet = "SiteToSheet.run:main"

[tool.setuptools]
packages = ["SiteToSheet"]